Fix: Refactor codes for Header.js file To improve the structure and apply Styled-components appropriately Ref: #134
